requirements.txt
datasets>=2.14.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
einops>=0.6.0
